+++ football-manager/src/main/java/com/example/football_manager/modules/team/repository/PlayerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PlayerRepository extends JpaRepository<Player, Long> {
-    List<Player> findByTeamId(Long teamId);
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
+    List<Player> findByTeamId(Integer teamId);
 }
