+++ football-manager/src/main/java/com/example/football_manager/modules/tournament/dto/AddTeamRequest.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 @Data
 public class AddTeamRequest {
-    private List<Long> teamIds;
+    private List<Integer> teamIds;
 }
